Create parsingdnslog.py
parsing DNS logs
dns_server - This VM will act as a router and DNS forwarder for your client VM. In other words all the Internet traffic from the client VM will go through dns_server VM. It will allow you to observe the traffic and monitor any DNS queries from the client. We preconfigured the bind9 on dns_server VM to log all the queries in /var/log/named/queries.log. This will be the file that you will use as an input to your program.
client - this vm is configured to use dns_server VM as a default router and DNS server. When you browse the web from this VM any DNS requests issued will be stored on the dns_server VM.